Only generate SQL password if none there already.

+++ ubuntu-automation/git-install-main.sh
@@ -15,9 +15,11 @@ if [ ! -f /var/newsreduce/network ] || [ ! "$(cat /var/newsreduce/network)" ]; t
 fi
 vi /var/newsreduce/network
 echo 1 > /var/newsreduce/is_main
-sql_password="$(dd if=/dev/urandom bs=15 count=1 2>/dev/null | base64 | tr '/' '_' | tr '+' ',')"
-echo "$sql_password" > /var/newsreduce/sql_password
-    echo Safely creating user...
+if [ ! -f /var/newsreduce/sql_password ]; then
+    sql_password="$(dd if=/dev/urandom bs=15 count=1 2>/dev/null | base64 | tr '/' '_' | tr '+' ',')"
+    echo "$sql_password" > /var/newsreduce/sql_password
+fi
+echo Safely creating user...
 echo "drop user if exists newsreduce;"                                                                                            | mysql
 echo "drop user if exists 'newsreduce'@'localhost';"                                                                              | mysql
 echo "drop user if exists 'newsreduce'@'%';"                                                                                      | mysql
